requirements_player_os.txt
# Player OS требуемые пакеты

# Дисплей
luma-lcd==2.13.0
luma-core==2.5.3

# Обработка изображений
Pillow>=10.0.0

# GPIO (обычно уже установлен на Raspberry Pi)
RPi.GPIO>=0.7.0

# Медиаплеер (система)
# ffmpeg - устанавливается через apt: sudo apt install ffmpeg

# Опционально для отладки
python-vlc>=3.0.0  # Если хотите вернуть VLC
